Connected servers to workers, Rebar integration
Added code for the gen_servers: cunit.erl, map_gen.erl and vehicle_data.erl

Also added code to the workers: car_ai.erl, cunit.erl, hardware_data.erl,
position_calc.erl

The different worker processes and the servers are able to pass information
to each other. Code is using dummy values, and TODO notations are added
to show what need to be cleaned up later on.

The gen_servers are added as supervisors, and the others as worker processes
in the rebar application. The application now runs all the processes when
run.sh is executed.